[MH3-115328] phy_tx_test_console pic32cx_mtg_ek_pl460 updated to new references

+++ apps/phy_apps/phy_tx_test_console/firmware/src/config/pic32cx_mtg_ek_pl460/service/pvddmon/srv_pvddmon.c
@@ -193,7 +193,7 @@ void SRV_PVDDMON_CallbackRegister (SRV_PVDDMON_CALLBACK callback, uintptr_t cont
 
 bool SRV_PVDDMON_CheckWindow(void)
 {
-    uint32_t adcValue;
+    uint16_t adcValue;
 
     adcValue = ADC_ChannelResultGet(ADC_CH4);
     while(adcValue == 0U)
